messages are now identified by uInt16 - less space used and 65,535 different ids - thus the chances of generating the same one is lower

+++ bluetoothChat/Models/ConversationModel.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct Conversation: Identifiable {
-    var id: Int
+    var id: UInt16
     var author: String
     var lastMessage: Message
     var messages: [Message]
